Added tests for repository
