+++ src/test/java/uk/ac/ebi/ampt2d/test/persistence/TestMonotonicEntity.java
@@ -18,7 +18,7 @@
 package uk.ac.ebi.ampt2d.test.persistence;
 
 import uk.ac.ebi.ampt2d.commons.accession.generators.ModelHashAccession;
-import uk.ac.ebi.ampt2d.commons.accession.persistence.jpa.accession.entity.AccessionedLongEntity;
+import uk.ac.ebi.ampt2d.commons.accession.persistence.jpa.accession.entities.AccessionedLongEntity;
 import uk.ac.ebi.ampt2d.test.TestModel;
 
 import javax.persistence.Entity;
